Add constants.h and bg_jobs.h
